create a firewall interface
This is creating a firewall interface and reintroduce the iptables integration as a fallback.
